+++ test/overwrites.spec.js
@@ -2,7 +2,7 @@ const index = require('./mock/server').server;
 const { aCommandSyn, Commands } = require('scure').dsl;
 
 const scure = buildScure();
-
+const THERE_IS_NO_ROOM_LIKE_CARTERA = 'Cannot read property';
 describe('SCURE - overwrite commands', () => {
   let oldConf;
   beforeEach(() => {
@@ -26,6 +26,35 @@ describe('SCURE - overwrite commands', () => {
     expect(getDfaV2Conv().lastAsk).to.contains('Ya me estás usando');
   });
 
+  it('Overwrites commands through configuration with arguments when name is an item', () => {
+    scure.data.commandSyns = scure.data.commandSyns || [];
+    scure.data.commandSyns.push(aCommandSyn(Commands.WALK, 'ric', Commands.USE));
+
+    const request = aDfaV2Request()
+      .withIntent('walk')
+      .withData({ roomId: 'comedor' })
+      .withArgs({ arg: 'robot' })
+      .build();
+
+    index.fulfillment(request);
+
+    expect(getDfaV2Conv().lastAsk).to.contains('Ya me estás usando');
+  });
+
+  it('does not overwrite commands when arguments not available', () => {
+    scure.data.commandSyns = scure.data.commandSyns || [];
+    scure.data.commandSyns.push(aCommandSyn(Commands.WALK, 'comedor-cartera', Commands.USE));
+
+    const request = aDfaV2Request()
+      .withIntent('walk')
+      .withData({ roomId: 'pasillo norte' })
+      .withArgs({ arg: 'cartera' })
+      .build();
+
+    expect(() => {
+      index.fulfillment(request);
+    }).to.throw(THERE_IS_NO_ROOM_LIKE_CARTERA);
+  });
 
 
   it('does not overwrite commands for different arguments', () => {